+++ packages/toolkit/src/lib/store/index.ts
@@ -1,6 +1,5 @@
 export * from "./useConfigureModelFormStore";
 export * from "./useConfigurePipelineFormStore";
-export * from "./useConfigureProfileFormStore";
 export * from "./useConfigureSourceFormStore";
 export * from "./useCreateResourceFormStore";
 export * from "./useMessageBoxStore";
